move status components to common components
