+++ Source/Applications/SystemCenter/wwwroot/Scripts/TSX/SystemCenter/CommonComponents/StatusGroup.tsx
