+++ Source/Applications/SystemCenter/wwwroot/Scripts/TSX/SystemCenter/CommonComponents/StatusItem.tsx
